% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{addLinkageGroup}
\alias{addLinkageGroup}
\title{Add linkage group to genetic system}
\usage{
addLinkageGroup(genopheno, nloci = 1, nalleles = rep(2, nloci),
  alleleNames = "standard", rec = NULL, type = "autosomal")
}
\arguments{
\item{genopheno}{genopheno object specifying the genetic setup.}

\item{nloci}{The number of loci. The default value is 1 but any (reasonable) number can be chosen.}

\item{nalleles}{The number of alleles for each locus. This should be a vector of length \code{nloci} containing positive integer numbers.
By default, all loci are set to be biallelic.}

\item{alleleNames}{The names for the alleles at the different loci.
The default value is 'standard', which for biallelic loci produces allele names 'a' and 'A' for the first locus, 'b' and 'B' for the second locus etc.
If one locus has more than two alleles, 'standard' produces allele names a0, a1, a2; b0, b1, b2 etc.
For other allele names, \code{alleleNames} should be a list of nloci elements, each of which needs to be a vector containing the allele names.
Allele names need to be strings (class \code{character}) and can contain letters, numbers,
or symbols (except '~', '.' and '|').}

\item{rec}{A vector of length \code{nloci-1} specifying recombination rates between the loci.}

\item{type}{This parameter specifies the ploidy and inheritance for the loci.
The default value is "autosomal" and currently, this is also the only supported type
of linkage group.}
}
\value{
An object of class \code{genopheno} that stores all of the specifications provided about the genetic setup
and that can then be extended to also include other linkage groups and corresponding phenotypes.
}
\description{
Add linkage group to genetic system
}
