% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertGenoStringToList}
\alias{convertGenoStringToList}
\title{Convert from genotype string to list format}
\usage{
convertGenoStringToList(genoString, genopheno)
}
\arguments{
\item{genoString}{Character string representing a genotype}

\item{genopheno}{The genetic system that the genotype is part of}
}
\value{
A genotype object in list format
}
\description{
Genotypes are internally represented as lists but specified as strings by the user.
This function converts from the latter into the former format.
}
\keyword{internal}
