% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createEmptyGenotype}
\alias{createEmptyGenotype}
\title{Creates an empty genotype in list format}
\usage{
createEmptyGenotype(genopheno)
}
\arguments{
\item{genopheno}{The genetic system that the genotype is part of}
}
\value{
A genotype object in list format
}
\description{
The generated genotype object has the specified number of linkage groups, loci etc.,
but only has NAs at all allelic positions.
}
\keyword{internal}
