% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Retrieve information about the phenotypes of the genetic system}
\usage{
getPhenotypes(genopheno, equivalent = "phase", hideNoncanonical = TRUE)
}
\arguments{
\item{genopheno}{A genopheno object containing all information about a genetic system.}

\item{equivalent}{Argument determining which genotypes are to be treated as equivalent.
Possible options are "phase" (genotypes that differ only by phase are equivalent,
i.e. the order of alleles within a locus does not matter),
"origin" (phase is important but not the origin of alleles),
and "none" (no two genotypes are equivalent).}
}
\value{
A dataframe in which rows represent genotypes (given by row names),
        and columns represent traits.
}
\description{
Retrieve information about the phenotypes of the genetic system
}
\examples{

MendelsPeas <- newGenopheno(nloci = 1, alleleNames = list(c('Y', 'y')))
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'Y_', 'yellow')
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'yy', 'green')
getPhenotypes(MendelsPeas)

}
