% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNoncanonical}
\alias{isNoncanonical}
\title{Checks if a genotype is a non-canonical genotype}
\usage{
isNoncanonical(gt, genopheno)
}
\arguments{
\item{gt}{a genotype in list format}

\item{genopheno}{a genopheno object containing all information about a genetic setup.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
With some types of linkage groups such as 'X-linked', there are genotypes that do not
usually arise (e.g., absence of both X chromosomes) and should therefore not be displayed by default.
This function checks if a genotype is such a non-canonical genotype.
}
\keyword{internal}
