% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchGenotypes}
\alias{matchGenotypes}
\title{Find the index of a genotype in a list of genotypes}
\usage{
matchGenotypes(gt, genotypes)
}
\arguments{
\item{gt}{Genotype (in list format)}

\item{genotypes}{List of genotypes}
}
\value{
Index of genotype within list, or \code{NA} if the genotype is not found
}
\description{
Find the index of a genotype in a list of genotypes
}
\keyword{internal}
