% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.genopheno}
\alias{print.genopheno}
\title{Print information about genotypes and corresponding phenotypes.}
\usage{
\method{print}{genopheno}(x, ...)
}
\arguments{
\item{x}{an obbject of class \code{genopheno}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print information about genotypes and corresponding phenotypes.
}
\examples{

MendelsPeas <- newGenopheno(nloci = 1, alleleNames = list(c('Y', 'y')))
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'Y_', 'yellow')
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'yy', 'green')
MendelsPeas

}
