% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setPhenotypes}
\alias{setPhenotypes}
\title{Set trait values for one or several genotypes.}
\usage{
setPhenotypes(genopheno, traitName, genotypes, traitValue,
  equivalent = "phase")
}
\arguments{
\item{genopheno}{genopheno object specifying the genetic setup.}

\item{traitName}{The name of the trait whose values are to be set.}

\item{genotypes}{A vector of strings specifying genotypes.}

\item{traitValue}{The trait value.}

\item{equivalent}{Argument determining which genotypes are to be treated as equivalent.
Possible options are "phase" (genotypes that differ only by phase are equivalent,
i.e. the order of alleles within a locus does not matter),
"origin" (phase is important but not the origin of alleles),
and "none" (no two genotypes are equivalent).}
}
\value{
genopheno object with trait information added.
}
\description{
Set trait values for one or several genotypes.
}
\examples{
MendelsPeas <- newGenopheno(nloci = 1, alleleNames = list(c('Y', 'y')))
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'Y_', 'yellow')
MendelsPeas <- setPhenotypes(MendelsPeas, 'colour', 'yy', 'green')
getPhenotypes(MendelsPeas)

}
