% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setSexPhenotypes}
\alias{setSexPhenotypes}
\title{Add sex phenotypes to a genetic system with sex chromosomes}
\usage{
setSexPhenotypes(genopheno)
}
\arguments{
\item{genopheno}{genopheno object specifying the genetic setup.}
}
\value{
genopheno object with sex trait information added.
}
\description{
This function add a trait 'sex' to an existing genopheno object.
Trait values ('female' and 'male') are assigned automatically to all genotypes
according to sex chromosome configurations.
}
\examples{
mySystem <- newGenopheno(type = "X-linked")
mySystem <- setSexPhenotypes(mySystem)
getPhenotypes(mySystem)

}
