% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{someButNotAllColumnZero}
\alias{someButNotAllColumnZero}
\title{Helper function to check for which rows of a dataframe some but not all columns are zero}
\usage{
someButNotAllColumnZero(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A vector with TRUE and FALSE elements
}
\description{
Helper function to check for which rows of a dataframe some but not all columns are zero
}
\keyword{internal}
