% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyGenoString}
\alias{tidyGenoString}
\title{Genotype string tidying}
\usage{
tidyGenoString(genoString)
}
\arguments{
\item{genoString}{The character string to be tidied.}
}
\value{
A tidy genotype string.
}
\description{
Converts a character string representing a genotype into a canonical form
where spaces within linkage groups are removed and linkage groups are
connected by ' | '.
}
\keyword{aA}
\keyword{internal}
\keyword{tidyGenoString("}
\keyword{~bb|CC")}
